remove duplicated parsing classes
