+++ src/main/java/com/mozilla/secops/httprequest/HTTPRequest.java
@@ -29,6 +29,7 @@ import com.mozilla.secops.parser.Event;
 import com.mozilla.secops.parser.EventFilter;
 import com.mozilla.secops.parser.EventFilterRule;
 import com.mozilla.secops.parser.Parser;
+import com.mozilla.secops.parser.ParserDoFn;
 import com.mozilla.secops.parser.Payload;
 import com.mozilla.secops.parser.GLB;
 
@@ -64,47 +65,12 @@ public class HTTPRequest implements Serializable {
 
         @Override
         public PCollection<Event> expand(PCollection<String> col) {
-            class Parse extends DoFn<String, Event> {
-                private static final long serialVersionUID = 1L;
-
-                private Logger log;
-                private Parser ep;
-                private Long parseCount;
-
-                @Setup
-                public void Setup() {
-                    ep = new Parser();
-                    log = LoggerFactory.getLogger(Parse.class);
-                    log.info("initialized new parser");
-                }
-
-                @StartBundle
-                public void StartBundle() {
-                    log.info("processing new bundle");
-                    parseCount = 0L;
-                }
-
-                @FinishBundle
-                public void FinishBundle() {
-                    log.info("{} events processed in bundle", parseCount);
-                }
-
-                @ProcessElement
-                public void processElement(ProcessContext c) {
-                    Event e = ep.parse(c.element());
-                    if (e != null) {
-                        parseCount++;
-                        c.output(e);
-                    }
-                }
-            }
-
             EventFilter filter = new EventFilter()
                 .setWantUTC(true)
                 .setOutputWithTimestamp(emitEventTimestamps);
             filter.addRule(new EventFilterRule().wantSubtype(Payload.PayloadType.GLB));
 
-            return col.apply(ParDo.of(new Parse()))
+            return col.apply(ParDo.of(new ParserDoFn()))
                 .apply(EventFilter.getTransform(filter))
                 .apply(Window.<Event>into(FixedWindows.of(Duration.standardMinutes(1))));
         }
